Removed edit functionality from here.

+++ src/app/components/article-full/article-full.component.ts
@@ -3,6 +3,9 @@ import { ArticlesService } from 'src/app/services/articles.service';
 import { ActivatedRoute, Router } from '@angular/router';
 import { IArticle } from 'src/app/models/entities/IArticle';
 
+import { MatDialog, MatDialogRef } from '@angular/material';
+import { DeleteArticleComponent } from '../modals/delete-article/delete-article.component';
+
 @Component({
   selector: 'app-article-full',
   templateUrl: './article-full.component.html',
@@ -10,45 +13,44 @@ import { IArticle } from 'src/app/models/entities/IArticle';
 })
 export class ArticleFullComponent implements OnInit {
   public article: IArticle;
-  public isEditMode: boolean;
-
-  private _articleBackup: IArticle;
 
   constructor(
+    public matDialog: MatDialog,
     private _articlesService: ArticlesService,
     private _route: ActivatedRoute,
     private _router: Router
   ) {
-    const id = this._route.snapshot.params.id;
-    this.article = this._articlesService.getArticle(id);
-    this._articleBackup = this._articlesService.getArticleCopy(this.article);
+    if (this._route.snapshot.params && this._route.snapshot.params.id) {
+      const id = this._route.snapshot.params.id;
+      this.article = this._articlesService.getArticle(id);
 
-    this.isEditMode = false;
+      if (!this.article)
+        this._router.navigate(['/not-found']);
+    }
+    else {
+      this._router.navigate(['/not-found']);
+    }
   }
 
   ngOnInit() {
   }
 
   public editArticle(): void {
-    this.isEditMode = true;
+    this._router.navigate([`/article/${this.article.id}/edit`]);
   }
 
   public deleteArticle(): void {
-    this._articlesService.deleteArticle(this.article.id);
-    this.isEditMode = false;
+    let dialogRef = this.matDialog.open(DeleteArticleComponent, {
+      height: '200px',
+      width: '350px',
+    });
 
-    this._router.navigate([""]);
-  }
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this._articlesService.deleteArticle(this.article.id);
 
-  public saveArticle(): void {
-    this._articlesService.updateArticle(this.article);
-    this._articleBackup = this._articlesService.getArticleCopy(this.article);
-
-    this.isEditMode = false;
-  }
-
-  public undoArticle(): void {
-    this.isEditMode = false;
-    this.article = this._articlesService.getArticleCopy(this._articleBackup);
+        this._router.navigate([""]);
+      }
+    });
   }
 }
